+++ benchmarks/thread_read_write.cpp
@@ -13,11 +13,11 @@ namespace po = boost::program_options;
 
 const std::string storage_path = "threadBenchmarkStorage";
 
-int meas2write = 10;
-int pagesize = 1000000;
-int thread_count = 5;
-int iteration_count = 1000000;
-int read_iteration_count=10;
+size_t meas2write = 10;
+size_t pagesize = 1000000;
+size_t thread_count = 5;
+size_t iteration_count = 1000000;
+size_t read_iteration_count=10;
 bool verbose = false;
 bool dont_remove = false;
 bool enable_dyn_cache = false;
@@ -57,7 +57,7 @@ void writer(int writeCount) {
 }
 
 void reader(int num,storage::Time from, storage::Time to) {
-	auto read_window = (to - from) / read_iteration_count;
+    auto read_window = (to - from) / read_iteration_count;
 
 	for (size_t i = 0; i < read_iteration_count; i++) {
 		auto reader = ds->readInterval(read_window*(i+1), read_window*(i+2));
@@ -107,12 +107,12 @@ void show_reads_info() {
 int main(int argc, char *argv[]) {
     po::options_description desc("Multithread IO benchmark.\n Allowed options");
 	desc.add_options()("help", "produce help message")(
-		"mc", po::value<int>(&meas2write)->default_value(meas2write), "measurment count")
+        "mc", po::value<size_t>(&meas2write)->default_value(meas2write), "measurment count")
 		("dyncache", po::value<bool>(&enable_dyn_cache)->default_value(enable_dyn_cache), "enable dynamic cache")
 		("cache-size", po::value<size_t>(&cache_size)->default_value(cache_size), "cache size")
 		("cache-pool-size", po::value<size_t>(&cache_pool_size)->default_value(cache_pool_size), "cache pool size")
-		("thread-count", po::value<int>(&thread_count)->default_value(thread_count), "write thread count")
-		("page-size", po::value<int>(&pagesize)->default_value(pagesize), "page size")
+        ("thread-count", po::value<size_t>(&thread_count)->default_value(thread_count), "write thread count")
+        ("page-size", po::value<size_t>(&pagesize)->default_value(pagesize), "page size")
 		("verbose", "verbose ouput")
 		("dont-remove", "dont remove created storage");
 
@@ -145,13 +145,13 @@ int main(int argc, char *argv[]) {
 	/// writers
 	std::vector<std::thread> writers(thread_count);
 	size_t pos = 0;
-	for (int i = 0; i < thread_count; i++) {
+    for (size_t i = 0; i < thread_count; i++) {
 		std::thread t{ writer, iteration_count };
 		writers[pos++] = std::move(t);
 	}
 
 	pos = 0;
-	for (int i = 0; i < thread_count; i++) {
+    for (size_t i = 0; i < thread_count; i++) {
 		std::thread t = std::move(writers[pos++]);
 		t.join();
 	}
@@ -165,13 +165,13 @@ int main(int argc, char *argv[]) {
 	std::vector<std::thread> readers(thread_count);
 	pos = 0;
 	auto reads_per_thread = (iteration_count / ((float)thread_count));
-	for (int i = 0; i < thread_count; i++) {
+    for (size_t i = 0; i < thread_count; i++) {
 		std::thread t{ reader,i, reads_per_thread*i, reads_per_thread*(i+1) };
 		readers[pos++] = std::move(t);
 	}
 
 	pos = 0;
-	for (int i = 0; i < thread_count; i++) {
+    for (size_t i = 0; i < thread_count; i++) {
 		std::thread t = std::move(readers[pos++]);
 		t.join();
 	}
